lgl22: fix lge touchscreen code
There was an error in touchscreen code as it relates to kddi (lgl22).

Line 4959:
`#if (defined(A1_only) && !defined(CONFIG_MACH_MSM8974_G2_KDDI)) ||
		defined(CONFIG_LGE_Z_TOUCHSCREEN)`

However, the other two portions of the code with the same condition,
did not actually have the same condition.  Compared code to official
lge source to determine which condition was the correct condition.
This commits makes all 3 conditions the same, as they should be, and
the same as the official lge source.